fix clippy lints
vec![] expands to Vec::new(), which is a non-allocating const function
and doesn't need a closure, but Vec also implements Default and that
triggered another lint to use unwrap_or_default() rather than
unwrap_or(vec![]).

The matches!() macro requires Rust >= 1.42.

Clippy suggested sort_unstable when sorting strings, which I guess is
fine, though it's not like the tests were slow anyway (we're sorting
a handful of strings in that code).