Don't forcibly add './' prefix when stdout isn't a terminal
The new behavior of fd is to always add './' when fd is piped anywhere,
and this messes with vimz (fz piping fd's output into sk) and things
like `fd ... | sort` or whatever.

I don't like this, but have to do some Crimes in order to avoid changing
a ton of test cases and create merge conflict hell for Future Me.

  * Change the normal behavior to ignore isatty and always strip the cwd
    prefix if no search path is given.
  * Add a secret environment variable FW_CWD_PREFIX_ON_PIPE which
    restores the current upstream behavior of fd. Update the testenv
    runner to set this environment variable during tests.
  * Hide the --strip-cwd-prefix option from help text, but leave it
    enabled for test_strip_cwd_prefix(). This option overrides the
    interactive terminal check when FD_CWD_PREFIX_ON_PIPE is set.