make clppy happy
Clippy now lints for "variables can be used directly in the `format!`
string" (uninlined_format_args), which is neat and can be fixed with
"cargo clippy --fix", but I don't want to deal with the merge conflicts
that'll inevitably create, so just allow the lint instead. But do
address the lint in my own additions to build.rs.