reduce MSRV to 1.74.0
Drop the rust-version down so that fd can build on Yocto Scarthgap,
which uses Rust 1.75.0. Rust 1.74.0 is required by clap_derive v4.5.13,
and tests pass with 1.74.0, so use that.

Upstream commit e262ade bumped rust-version to 1.77.2 for the sake of
fixes for windows cmd.exe argument escaping insanity (CVE-2024-24576)
but the rust-version of a package should really be what's *needed*, not
what maintainers think users should be using.