deps: replace 'etcetera' with 'dirs'
A couple years ago (3ae0454), 'dirs-next' was replaced with 'etcetera',
which depends on the 'home' crate that's maintained as part of cargo (in
the cargo monorepo). Because 'home' is part of cargo, it has a very
aggressive MSRV, currently at 1.81, and this filters up to fd.  Yocto
styhead uses Rust 1.79, so that's a problem.

The reasons for moving away from dirs(-next) seem to be more political
than technical - it seems some people *really* dislike the MPL, or
object to the option-ext crate even though it's super tiny.

On the bright side, the OG 'dirs' is maintained again (unlike
'dirs-next') so switch to that and verify that we can build with 1.79.